Fill in correct preset loop options on sample load
People have been sharing copy-paste-ready URLs with the !o,p,t,i,o,n,s!
syntax out there, to be used via "Add multiple samples".

When using the loop-control-related options (a, b, c, d, and e), if the
loop start/end were not filled in, they would previously default to
[0, 2], which is incorrect and useless.

Now, they should still default to [0, 2], but if the sample finishes
loading, the values will be rewritten to [0, sample.length - 1], if not
specified.

This is potentially confusing, but ideally the user should wait for the
samples to load anyway, as some other functionality is also disabled
when they are still loading (like the loop controls prompt).