Only compute loop control data for chip wave tones
Rather serious oversight. Without this guard, the code could end up
trying to load the length of a waveform that isn't present in Config
(and thus undefined).

The problem was commonly seen when loading a song that uses custom
samples, and then making a new blank song (the custom sample song would
need to use a sample on channel 4 in order for this to be reproduced).